% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_quota.R
\name{adjust_quota}
\alias{adjust_quota}
\title{Adjust API Quota for a Given Token}
\usage{
adjust_quota(token, schema = "adem")
}
\arguments{
\item{token}{A character string representing the API token of the user.}

\item{schema}{A character string specifying the schema name in the PostgreSQL database. Default is `'adem'`.}
}
\value{
Returns `TRUE` if the quota was successfully updated.
}
\description{
Decreases the quota by 1 for the user identified by a given API token.
Ensures the quota does not fall below 0 and throws an error if the token is not found or the quota is already exhausted.
}
\examples{
\dontrun{
adjust_quota("abc123")
adjust_quota("xyz789", schema = "public")
}
}
