% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{connect_db}
\alias{connect_db}
\title{Connect to Database}
\usage{
connect_db()

connect_db()
}
\value{
A DBI connection object (class \code{PqConnection})
}
\description{
Establishes a connection to the PostgreSQL database using credentials
and host information stored in environment variables.
}
\examples{
\dontrun{
con <- connect_db()
DBI::dbListTables(con)
DBI::dbDisconnect(con)
}
}
