% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_users.R
\name{create_user}
\alias{create_user}
\title{Create a new API user}
\usage{
create_user(username, token, quota = 100)
}
\arguments{
\item{username}{The new username}

\item{token}{The associated API token}

\item{quota}{Number of requests allowed (default = 100)}
}
\value{
TRUE if the user was created, FALSE otherwise
}
\description{
Create a new API user
}
