% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_aut.R
\name{decrement_quota}
\alias{decrement_quota}
\title{Decrements API call user quota (given a validated token)}
\usage{
decrement_quota(conn, token, schema)
}
\arguments{
\item{conn}{A valid DB connection object.}

\item{token}{The validated user token (a string).}

\item{schema}{The DB schema to use for user lookup/update.}
}
\value{
Nothing.
}
\description{
Decrements API call user quota (given a validated token)
}
