% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect_db.R
\name{env_var_set_not_empty}
\alias{env_var_set_not_empty}
\title{Checks if an environment variable is set (in .Renviron) but is not an empty string (if set, it will always be a string).}
\usage{
env_var_set_not_empty(env_var_name)
}
\arguments{
\item{env_var_name}{Name of environment variable to be checked}
}
\value{
TRUE, if the checks are successful; FALSE, in case any of the criteria are not matched.
}
\description{
Checks if an environment variable is set (in .Renviron) but is not an empty string (if set, it will always be a string).
}
\examples{
\dontrun{
env_var_set_not_empty("PG_DB")
env_var_set_not_empty("any-variable-name")
}
}
