% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_recommendations.R
\name{get_book}
\alias{get_book}
\title{Get Book Recommendations}
\usage{
get_book(param1)
}
\arguments{
\item{param1}{}
}
\value{
A data frame with columns 'book_id', 'title', 'author', and 'skill_id'.
}
\description{
Returns a list of recommended books. You can filter them by skill.
}
\examples{
\dontrun{
getbooks()
}
}
