% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_companies_details.R
\name{get_companies_details}
\alias{get_companies_details}
\title{Retrieve Company Information and Vacancies}
\usage{
get_companies_details(company_id = 1)
}
\arguments{
\item{company_id}{Integer. A unique identifier for the company whose details and vacancies
are to be retrieved. Defaults to 1.}
}
\value{
A data frame containing the company information (including the column
\code{company_name}) along with all its vacancies. If the company is not found, the function
returns \code{NULL}.
}
\description{
This function connects to the \code{adem} database and retrieves detailed information
for a specified company. It returns both the company information and all associated
vacancies by performing a join between the \code{adem.companies} and \code{adem.vacancies} tables.
If the company is not found, the function returns \code{NULL}.
}
\examples{
\dontrun{
  # Retrieve info and vacancies for the company with company_id 1
  get_companies_details(1)

  # Attempt to retrieve info for a non-existing company; returns NULL
  get_companies_details(company_id = 1)
}

}
\seealso{
\code{\link[=connect_db]{connect_db()}}, \code{\link[DBI:dbGetQuery]{DBI::dbGetQuery()}}, \code{\link[glue:glue_sql]{glue::glue_sql()}}
}
