% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_companies_details.R
\name{get_company_details}
\alias{get_company_details}
\title{Get a book by its ID}
\usage{
get_company_details(company_id)
}
\arguments{
\item{book_id}{The unique identifier of the book.}
}
\value{
A data frame with columns: book_id, title, author, skill_id. Returns NULL if no book is found.
}
\description{
Get a book by its ID
}
