% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_learning_track_by_id.R
\name{get_learning_track_by_id}
\alias{get_learning_track_by_id}
\title{Get a book by its ID}
\usage{
get_learning_track_by_id(track_id)
}
\arguments{
\item{book_id}{The unique identifier of the book.}
}
\value{
A data frame with columns: book_id, title, author, skill_id. Returns NULL if no book is found.
}
\description{
Get a book by its ID
}
