% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skills.R
\name{get_skill}
\alias{get_skill}
\title{Get a book by its ID}
\usage{
get_skill()
}
\arguments{
\item{book_id}{The unique identifier of the book.}
}
\value{
A data frame with columns: book_id, title, author, skill_id. Returns NULL if no book is found.
}
\description{
Get a book by its ID
}
