% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skills.R
\name{get_skills}
\alias{get_skills}
\title{Returns a list of registered skills from the database.
Using the function without the limit parameter the function will return all registered skills from the database.}
\usage{
get_skills(limit = 100)
}
\arguments{
\item{limit}{A numeric type that limits the maximum number of results.}
}
\value{
A data frame with a list of skills.
}
\description{
Returns a list of registered skills from the database.
Using the function without the limit parameter the function will return all registered skills from the database.
}
\examples{
\dontrun{
get_skills()
get_skills(12)
get_skills(limit = 20)
}
}
