% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vacancies.R
\name{get_vacancies}
\alias{get_vacancies}
\title{Returns a list of job vacancies.
Optionally filters can be given as parameters. These optional parameters:
skill -> a skill in character format
company -> a name of a company
canton -> the job location
limit -> limit for the size of search results}
\usage{
get_vacancies(skill = NULL, company = NULL, canton = NULL, limit = 100)
}
\arguments{
\item{skill}{The name of the skill.}

\item{company}{The ID of the company.}

\item{canton}{The name of a canton.}

\item{limit}{the maximum number of results from the query.}
}
\value{
A data frame with the list of job vacancies.
}
\description{
Returns a list of job vacancies.
Optionally filters can be given as parameters. These optional parameters:
skill -> a skill in character format
company -> a name of a company
canton -> the job location
limit -> limit for the size of search results
}
\examples{
\dontrun{
get_vacancies()
get_vacancies(skill = "show initiative")
get_vacancies(canton = "Mersch")
get_vacancies(skill = "show initiative", canton = "Mersch", limit = 3)
get_vacancies(canton = "Mersch", limit = 12)
}
}
