% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacancies.R
\name{get_vacancies_with_skill}
\alias{get_vacancies_with_skill}
\title{Get Vancancies and their skills}
\usage{
get_vacancies_with_skill(
  skill_id = NULL,
  company_id = NULL,
  canton = NULL,
  limit = 100
)
}
\arguments{
\item{skill_id}{Character.}

\item{company_id}{Integer. Optional filter for company_id.}

\item{canton}{Character. Optional filter for canton in Luxembourg.}

\item{limit}{Integer. The maximum number of vacancies to return.}
}
\value{
A data frame with the columns vacancy_id, company_id, occupation, canton, year, month and skill_id.
}
\description{
Returns the list of vacancies. You can filter for skill_id, company_id and/or canton. You can also limit the number of rows.
}
\examples{
\dontrun{
get_vacancies_with_skill()

get_vacancies_with_skill(canton="Luxembourg", limit = 10)
}
}
