% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vacancy_by_id.R
\name{get_vacancy_by_id}
\alias{get_vacancy_by_id}
\title{Returns full vacancy info and the required skills for that particular job vacancy.}
\usage{
get_vacancy_by_id(vacancy_id = NULL)
}
\arguments{
\item{vacancy_id}{A numeric type ID of a vacancy.}
}
\value{
A list type value with the vacancy info and the required skills for that particular job vacancy.
}
\description{
Returns full vacancy info and the required skills for that particular job vacancy.
}
\examples{
\dontrun{
get_vacancy_by_id(45)
get_vacancy_by_id(vacancy_id = 45)
}
}
