% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_aut.R
\name{verify_token}
\alias{verify_token}
\title{Validates a user token via DB (and decreases user API call quota, if token validated)}
\usage{
verify_token(token, schema)
}
\arguments{
\item{token}{The string representing the Bearer Token to be used for user authentication/authorization.}

\item{schema}{The name of the schema (a string) to be used for looking up API users.}
}
\value{
TRUE, if token has been found in DB (i.e., it belongs to one existing user); FALSE, otherwise.
}
\description{
Validates a user token via DB (and decreases user API call quota, if token validated)
}
\examples{
\dontrun{
verify_token("token123", "student_miguel")
}
}
