% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_quota.R
\name{adjust_quota}
\alias{adjust_quota}
\title{Adjust API Quota}
\usage{
adjust_quota(token)
}
\arguments{
\item{token}{Character Representing the API token of the user.}
}
\value{
Returns \code{TRUE} if the quota was successfully updated.
}
\description{
Decreases the quota by 1 for the user identified by a given API token.
Ensures the quota does not fall below 0 and throws an error if the token is not found or the quota is already exhausted.
}
\examples{
\dontrun{
adjust_quota("abc123")
}
}
