% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auth_aut.R
\name{decrement_quota}
\alias{decrement_quota}
\title{Decrements API call user quota (given a validated token), if quota > 0}
\usage{
decrement_quota(conn, token, schema)
}
\arguments{
\item{conn}{A valid DB connection object.}

\item{token}{The validated user token (a string).}

\item{schema}{The DB schema to use for user lookup/update.}
}
\value{
TRUE, if user quota was > 0; FALSE, if user quota was already zero.
}
\description{
Decrements API call user quota (given a validated token), if quota > 0
}
\examples{
\dontrun{
decrement_quota(conn, "token456", "student_miguel")
}
}
