% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_book_by_id.R
\name{get_book_by_id}
\alias{get_book_by_id}
\title{Returns a list of recommended books.
The parameter is a numeric type that represents a book ID number.
Calling the function without parameter will return a data frame with all recommended books.}
\usage{
get_book_by_id(book_id = NULL)
}
\arguments{
\item{book_id}{A numeric type data that represents the ID of a book.}
}
\value{
A data frame with the recommended books.
}
\description{
Returns a list of recommended books.
The parameter is a numeric type that represents a book ID number.
Calling the function without parameter will return a data frame with all recommended books.
}
\examples{
\dontrun{
get_book_by_id()
get_book_by_id(12)
get_book_by_id(book_id = 34)
}
}
