% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\name{get_book_by_id}
\alias{get_book_by_id}
\title{Obtains the details of a book, given its ID.}
\usage{
get_book_by_id(book_id = NULL)
}
\arguments{
\item{book_id}{The ID of the book whose details are to be retrieved.}
}
\value{
A data frame containing the book details. In case an invalid or missing book ID is provided, the function provides the details of book ID 1.
}
\description{
Obtains the details of a book, given its ID.
}
\examples{
\dontrun{
get_book_by_id(101)
}
}
