% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_book_by_id.R
\name{get_book_by_id}
\alias{get_book_by_id}
\title{get_book_by_id}
\usage{
get_book_by_id(book_id)
}
\arguments{
\item{book_id:}{integer}
}
\value{
data.frame with one row
}
\description{
Returns a list of recommended books. You can filter them by skill (optional)
}
\examples{
\dontrun{
get_book_by_id(101)
}


}
