% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_recommendations.R
\name{get_books}
\alias{get_books}
\title{Get Book Recommendations}
\usage{
get_books(skill = NULL)
}
\arguments{
\item{skill}{character. Optional filter the book recommendations for skill_label.}
}
\value{
A data frame with columns 'book_id', 'title', 'author', and 'skill_id'.
}
\description{
Returns a list of recommended books. You can filter them by skill.
}
\examples{
\dontrun{
#Get all books
get_books()

#Get books with specified skill
get_books("communication")
}
}
