% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_recommendations.R
\name{get_books}
\alias{get_books}
\title{Get books}
\usage{
get_books(skill = NULL)
}
\arguments{
\item{skill}{Character. Optional. Filter books by skill ID.}
}
\value{
A data frame with columns: \code{book_id}, \code{title}, \code{author}, \code{skill_id}.
}
\description{
Returns recommended books. You can optionally filter by a skill.
}
\examples{
\dontrun{
  get_books()
  get_books(skill = "skill_python")
}
}
