% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_books.R
\name{get_books}
\alias{get_books}
\title{Returns all recommended books.
The parameter is a character type that represents a skill label.
Calling the function without parameter will return a data frame with all recommended books.}
\usage{
get_books(skill = NULL)
}
\arguments{
\item{skill}{A string that represents a skill.}
}
\value{
A data frame with info about the recommended books.
}
\description{
Returns all recommended books.
The parameter is a character type that represents a skill label.
Calling the function without parameter will return a data frame with all recommended books.
}
\examples{
\dontrun{
get_books()
get_books("show initiative")
get_books(skill = "show initiative")
}
}
