% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\name{get_books}
\alias{get_books}
\title{Obtain list of the books, optionally filtered by a skill name.}
\usage{
get_books(skill = NULL)
}
\arguments{
\item{skill}{The skill name (label) (a string) describing the skill.}
}
\value{
A data frame with the details of all books, or otherwise, if a skill name (even if partial) is provided, then it returns book details matching that skill.
}
\description{
Obtain list of the books, optionally filtered by a skill name.
}
\examples{
\dontrun{
get_books(skill = "skill_python")
}
}
