% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/book_recommendations.R
\name{get_books}
\alias{get_books}
\title{Get Book Recommendations}
\usage{
get_books(skill = NULL)
}
\arguments{
\item{skill}{Optional. Character. A `skill_id` to filter recommendations.}
}
\value{
A data frame with columns `book_id`, `title`, `author`, and `skill_id`.
}
\description{
Returns a list of recommended books. You can filter them by skill.
}
\examples{
\dontrun{
get_books()
get_books(skill = "http://data.europa.eu/esco/skill/70198e4e-86ad-4acc-a9eb-e24e2c107d18")
}
}
