% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_books.R
\name{get_books}
\alias{get_books}
\title{get_books}
\usage{
get_books(skill = NULL)
}
\arguments{
\item{skill}{character (optional)}
}
\value{
data.frame with book_id, title, author, skill_id (replaced by label)
}
\description{
Returns all recommended books. Filter by skill if needed.
}
\examples{
\dontrun{
get_books(skill = "skill_python")
}
}
