% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies.R
\name{get_companies}
\alias{get_companies}
\title{Get list of companies}
\usage{
get_companies(limit = 100)
}
\arguments{
\item{limit}{}
}
\value{
A dataframe with columns 'company_id', 'name' and 'sector' with limited number of rows.
}
\description{
Get list of companies
}
\examples{
\dontrun{
get_companies(10)
}
}
