% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_companies.R
\name{get_companies}
\alias{get_companies}
\title{Returns all companies listed in the database.
Optionally the limit parameter can be set to control the number of results.
Calling the function without parameter will return the first 100 companies.}
\usage{
get_companies(limit = 100)
}
\arguments{
\item{limit}{A numeric data type to limit the number of results.}
}
\value{
A data frame with all companies.
}
\description{
Returns all companies listed in the database.
Optionally the limit parameter can be set to control the number of results.
Calling the function without parameter will return the first 100 companies.
}
\examples{
\dontrun{
get_companies()
get_companies(200)
get_companies(limit = 20)
}
}
