% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies.R
\name{get_companies}
\alias{get_companies}
\title{Obtains a list of the companies.}
\usage{
get_companies(limit = 100)
}
\arguments{
\item{limit}{The size (an integer) of the list of companies to be returned.}
}
\value{
A data frame containing the list of the companies and corresponding details. List can be limited using limit argument.
}
\description{
Obtains a list of the companies.
}
\examples{
\dontrun{
get_companies(limit = 50)
}
}
