% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_companies.R
\name{get_companies}
\alias{get_companies}
\title{get_companies}
\usage{
get_companies(limit = 100)
}
\arguments{
\item{limit}{integer}
}
\value{
data.frame with company_id, name, sector
}
\description{
Returns all companies from adem.companies.
}
\examples{
\dontrun{
get_companies(limit = 50)
}
}
