% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies.R
\name{get_company_details}
\alias{get_company_details}
\title{Get company and its vacancies}
\usage{
get_company_details(company_id)
}
\arguments{
\item{company_id}{Integer. Filter the results for company_id.}
}
\value{
A list with two data frames, the first one contains the company details and the second one contains the details of the vacancies associated with that company.
}
\description{
Returns the company selected by its ID and the vacancies related to it.
}
\examples{
\dontrun{
get_company_details(10)
}
}
