% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/companies.R
\name{get_company_details}
\alias{get_company_details}
\title{Get company and its vacancies}
\usage{
get_company_details(company_id)
}
\arguments{
\item{company_id}{}
}
\value{
A list with two data frames, the first one contains the company details and the second contains the details of the vacancies associated with that company.
}
\description{
Get company and its vacancies
}
\examples{
#' \dontrun{
get_company_details(10)
}
}
