% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_details.R
\name{get_company_details}
\alias{get_company_details}
\title{Returns all company information and all the vacancies from that company.
If no parameter is given, it returns NULL.}
\usage{
get_company_details(company_id = NULL)
}
\arguments{
\item{company_id}{A numeric type parameter that represents a company ID.}
}
\value{
A list data type with all company info and the related job vacancies.
}
\description{
Returns all company information and all the vacancies from that company.
If no parameter is given, it returns NULL.
}
\examples{
\dontrun{
get_company_details()
get_company_details(12)
}
}
