% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_details.R
\name{get_company_details}
\alias{get_company_details}
\title{Title}
\usage{
get_company_details(company_id)
}
\arguments{
\item{company_id}{}
}
\description{
Title
}
\examples{
\dontrun{
get_company_details(42)
}
}
