% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_tracks.R
\name{get_learning_track_by_id}
\alias{get_learning_track_by_id}
\title{Get learning track and its skills}
\usage{
get_learning_track_by_id(track_id)
}
\arguments{
\item{track_id}{}
}
\value{
A list with two data frames, the first one contains the learning track details and the second one contains the details of the skills associated with that learning track
}
\description{
Get learning track and its skills
}
\examples{
\dontrun{
get_learning_track_by_id(10)
}
}
