% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_learning_track_by_id.R
\name{get_learning_track_by_id}
\alias{get_learning_track_by_id}
\title{Returns one learning track and all its linked skills as a list.
If no parameter is given, then it returns NULL.}
\usage{
get_learning_track_by_id(track_id = NULL)
}
\arguments{
\item{track_id}{The ID of an ADEM learning track (numeric type).}
}
\value{
A list with a learning track with its linked skills.
}
\description{
Returns one learning track and all its linked skills as a list.
If no parameter is given, then it returns NULL.
}
\examples{
\dontrun{
get_learning_track_by_id()
get_learning_track_by_id(28)
get_learning_track_by_id(track_id = 28)
}
}
