% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_tracks.R
\name{get_learning_track_by_id}
\alias{get_learning_track_by_id}
\title{Obtains the details of a learning track, given its ID, and its associated skills.}
\usage{
get_learning_track_by_id(track_id = NULL)
}
\arguments{
\item{track_id}{The ID of the learning track whose details are to be retrieved.}
}
\value{
A list of 2 data frames: first, contains the details of the learning track; second, contains the skils (details) associated to it. In case an invalid track ID is provided, the function returns data about track ID 1. If the track ID provided is not found, the function returns NULL.
}
\description{
Obtains the details of a learning track, given its ID, and its associated skills.
}
\examples{
\dontrun{
get_learning_track_by_id(71)
}
}
