% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_learning_track_by_id.R
\name{get_learning_track_by_id}
\alias{get_learning_track_by_id}
\title{get_learning_track_by_id}
\usage{
get_learning_track_by_id(track_id)
}
\arguments{
\item{track_id}{integer}
}
\value{
list(track, skills)
}
\description{
Returns one learning track and all its linked skills. Returns NULL if not found.
}
\examples{
\dontrun{
get_learning_track_by_id(71)
}
}
