% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_learning_track.R
\name{get_learning_tracks}
\alias{get_learning_tracks}
\title{Get learning tracks}
\usage{
get_learning_tracks(skill_id = NULL)
}
\arguments{
\item{skill_id}{The unique identifier of the skill to filter tracks by.}
}
\value{
A data frame with columns: track_id, title, description, url. Returns all tracks if no skill_id is provided.
}
\description{
Get learning tracks
}
