% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_tracks.R
\name{get_learning_tracks}
\alias{get_learning_tracks}
\title{Get Learning Tracks}
\usage{
get_learning_tracks(skill_id = NULL)
}
\arguments{
\item{skill_id}{Character. Optional filter the learning tracks for skill_id.}
}
\value{
A data frame with columns 'track_id', 'title', 'description' and 'url'.
}
\description{
Returns a list of learning tracks.
}
\examples{
\dontrun{
#Get all learning tracks
get_learning_tracks()

#Get learning tracks with specified ESCO skill
get_learning_tracks("http://data.europa.eu/esco/skill/d8903406-abc4-48be-9b2e-5d8ddf103bd9")
}
}
