% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_learning_tracks.R
\name{get_learning_tracks}
\alias{get_learning_tracks}
\title{Returns learning tracks. If the skill_id parameter is given, then it returns only tracks teaching that particular skill.}
\usage{
get_learning_tracks(skill = NULL)
}
\arguments{
\item{skill}{A character type parameter representing a skill.}
}
\value{
A data frame with a learning track data.
}
\description{
Returns learning tracks. If the skill_id parameter is given, then it returns only tracks teaching that particular skill.
}
\examples{
\dontrun{
get_learning_tracks()
get_learning_tracks("show initiative")
get_learning_tracks(skill = "show initiative")
}
}
