% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/learning_tracks.R
\name{get_learning_tracks}
\alias{get_learning_tracks}
\title{Obtains a list of the learning tracks, optionally filtered by a skill ID}
\usage{
get_learning_tracks(skill_id = NULL)
}
\arguments{
\item{skill_id}{The ID (a string) of the skill to be used as a filter.}
}
\value{
A data frame containing a list of all learning tracks, or, in case an (existing) skill ID is provided, it returns only those learning tracks matching that skill.
}
\description{
Obtains a list of the learning tracks, optionally filtered by a skill ID
}
\examples{
\dontrun{
get_learning_tracks(skill_id = "skill_r")
}
}
