% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skill_by_id.R
\name{get_skill_by_id}
\alias{get_skill_by_id}
\title{Returns a single skill based on its ID.}
\usage{
get_skill_by_id(skill_id = NULL)
}
\arguments{
\item{skill_id}{The ID of a skill (character type).}
}
\value{
A data frame with a skill information.
}
\description{
Returns a single skill based on its ID.
}
\examples{
\dontrun{
get_skill_by_id(12)
get_skill_by_id(skill_id = 34)
}
}
