% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skill_by_id.R
\name{get_skill_by_id}
\alias{get_skill_by_id}
\title{Retrieve Skill Information by ESCO Skill ID}
\usage{
get_skill_by_id(skill_id)
}
\arguments{
\item{skill_id}{A character string containing the full ESCO Skill URI.
For example: \code{"http://data.europa.eu/esco/skill/000f1d3d-220f-4789-9c0a-cc742521fb02"}.}
}
\value{
A data frame containing the information of the skill corresponding to
the specified \code{skill_id}. If the ID does not exist in the database, an empty data frame is returned.
}
\description{
This function connects to the database and retrieves information from the
\code{adem.skills} table for a given ESCO Skill ID. The ID should be provided as a full URI string,
for example: \code{"http://data.europa.eu/esco/skill/000f1d3d-220f-4789-9c0a-cc742521fb02"}.
}
\examples{
# Retrieve a skill using its ESCO Skill URI
get_skill_by_id("http://data.europa.eu/esco/skill/000f1d3d-220f-4789-9c0a-cc742521fb02")

}
