% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills.R
\name{get_skill_by_id}
\alias{get_skill_by_id}
\title{Obtains the details of a skill, given its ID}
\usage{
get_skill_by_id(skill_id = NULL)
}
\arguments{
\item{skill_id}{The ID of skill (a string) whose details are to be retrieved.}
}
\value{
A data frame containing the skill details. If no skill ID or an invalid skill ID is provided, then it returns the details about the Communication skill.
}
\description{
Obtains the details of a skill, given its ID
}
\examples{
\dontrun{
get_skill_by_id("skill_r")
}
}
