% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills.R
\name{get_skill_by_id}
\alias{get_skill_by_id}
\title{Get Skill by ID}
\usage{
get_skill_by_id(skill_id)
}
\arguments{
\item{skill_id}{A character string. The ID of the skill to retrieve.}
}
\value{
A data frame with one row and columns `skill_id`, `skill_label`, or an empty data frame if not found.
}
\description{
Returns a single skill based on its ID from the `adem.skills` table.
}
\examples{
\dontrun{
get_skill_by_id("http://data.europa.eu/esco/skill/d8903406-abc4-48be-9b2e-5d8ddf103bd9")
}
}
