% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skill_by_id.R
\name{get_skill_by_id}
\alias{get_skill_by_id}
\title{get_skill_by_id}
\usage{
get_skill_by_id(skill_id)
}
\arguments{
\item{skill_id}{character}
}
\value{
data.frame with skill_id, skill_label
}
\description{
Returns a single skill based on its ID.
}
\examples{
\dontrun{
get_skill_by_id("skill_r")
}
}
