% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills.R
\name{get_skills}
\alias{get_skills}
\title{Get list of skills}
\usage{
get_skills(limit = 100, ordered = c("no", "yes"))
}
\arguments{
\item{limit}{Integer. The maximum number of skills to return.}

\item{ordered}{Character. Optional select if the list will be sorted or not.}
}
\value{
A dataframe with columns 'skill_id' and 'skill_label', limited number of rows and alphabetically ordered or not.
}
\description{
Returns a list of skills. You can limit the number of rows and order the skills alphabetically.
}
\examples{
\dontrun{
#Get list of skills with 10 rows and not ordered
get_skills(10)

#Get list of skills with 100 rows alphabetically ordered
get_skills(100, "yes")
}
}
