% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills.R
\name{get_skills}
\alias{get_skills}
\title{Get all skills}
\usage{
get_skills(limit = 100)
}
\arguments{
\item{limit}{Integer. Maximum number of skills to return. Default is 100.}
}
\value{
A data frame with columns: \code{skill_id}, \code{skill_label}.
}
\description{
Returns a list of skills from the \code{adem.skills} table, limited by \code{limit}.
}
\examples{
\dontrun{
  get_skills(limit = 20)
}
}
