% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skills.R
\name{get_skills}
\alias{get_skills}
\title{Obtain a list of the skills.}
\usage{
get_skills(limit = 100)
}
\arguments{
\item{limit}{The size (an integer) of the list of skills to be returned.}
}
\value{
A data frame with the list of the skills and corresponding details. List can be limited using limit argument.
}
\description{
Obtain a list of the skills.
}
\examples{
\dontrun{
get_skills(limit = 20)
}
}
