% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_skills.R
\name{get_skills}
\alias{get_skills}
\title{get_skills}
\usage{
get_skills(limit = 100)
}
\arguments{
\item{limit}{integer}
}
\value{
data.frame with skill_id, skill_label
}
\description{
Returns all skills from adem.skills. Limited by limit.
}
\examples{
\dontrun{
get_skills(limit = 20)
}
}
