% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacancies.R
\name{get_vacancies}
\alias{get_vacancies}
\title{Get vacancies}
\usage{
get_vacancies(occupation = NULL, company_id = NULL, canton = NULL, limit = 100)
}
\arguments{
\item{occupation}{Character. Optional filter for the occupation (job title).}

\item{company_id}{Integer. Optional filter for company_id.}

\item{canton}{Character. Optional filter for canton in Luxembourg.}

\item{limit}{Integer. The maximum number of vacancies to return.}
}
\value{
A data frame with the columns vacancy_id, company_id, occupation, canton, year and month.
}
\description{
Returns the list of vacancies. You can filter for occupation (job title), company_id and/or canton. You can also limit the number of rows.
}
\examples{
\dontrun{
get_vacancies()

get_vacancies(canton="Luxembourg")
}
}
