% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacancies.R
\name{get_vacancies}
\alias{get_vacancies}
\title{Get Job Vacancies}
\usage{
get_vacancies(skill = NULL, company = NULL, canton = NULL, limit = 100)
}
\arguments{
\item{skill}{Optional. Character. A `skill_id` to filter vacancies by required skill.}

\item{company}{Optional. Integer. A `company_id` to filter vacancies by company.}

\item{canton}{Optional. Character. A canton name to filter vacancies by location.}

\item{limit}{Optional. Integer. Max number of results (default = 100).}
}
\value{
A data frame of vacancies including `vacancy_id`, `company_id`, `occupation`, `canton`, `year`, and `month`.
}
\description{
Returns a filtered list of job vacancies. You can filter by skill, company, and/or canton.
}
\examples{
\dontrun{
get_vacancies()
get_vacancies(skill = "http://data.europa.eu/esco/skill/d8903406-abc4-48be-9b2e-5d8ddf103bd9", canton = "Luxembourg", limit = 50)
}
}
