% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vacancies.R
\name{get_vacancies}
\alias{get_vacancies}
\title{Title}
\usage{
get_vacancies(skill = NULL, company = NULL, canton = NULL, limit = 100)
}
\arguments{
\item{limit}{}
}
\description{
Title
}
\examples{
\dontrun{
get_vacancies(skill = "skill_python", canton = "Luxembourg")
}
}
