% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacancies.R
\name{get_vacancy_by_id}
\alias{get_vacancy_by_id}
\title{Get vacancy and its skills}
\usage{
get_vacancy_by_id(vacancy_id)
}
\arguments{
\item{vacancy_id}{Integer. Filter for the vancancy_id.}
}
\value{
A list with two data frames, the first one contains the vacancy details and the second one contains the skills associated with that vacancy.
}
\description{
Returns the selected vacancy and the skills connected to it.
}
\examples{
\dontrun{
get_vacancy_by_id(917152540)
}
}
