% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacancies.R
\name{get_vacancy_by_id}
\alias{get_vacancy_by_id}
\title{Get vacancy by ID}
\usage{
get_vacancy_by_id(vacancy_id)
}
\arguments{
\item{vacancy_id}{Integer. The ID of the vacancy.}
}
\value{
A list with two elements: \code{vacancy} (data frame with one row) and \code{skills} (data frame with 0 or more rows).
Returns \code{NULL} if the vacancy is not found.
}
\description{
Returns a vacancy and its required skills based on the given \code{vacancy_id}.
}
\examples{
\dontrun{
  get_vacancy_by_id(123456)
}
}
