% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vacancy_by_id.R
\name{get_vacancy_by_id}
\alias{get_vacancy_by_id}
\title{Get Vacancy Information by ID}
\usage{
get_vacancy_by_id(vacancy_id = NULL)
}
\arguments{
\item{vacancy_id}{A numeric or character value representing the unique identifier of the vacancy.
If \code{NULL}, all vacancies are returned (default is \code{NULL}).}
}
\value{
A data frame containing the vacancy details, including:
\itemize{
\item vacancy_id
\item vacancy_title
\item company_name
\item skill_label
}
If no vacancy is found for the provided \code{vacancy_id}, \code{NULL} is returned.
}
\description{
This function retrieves detailed vacancy information, including the related company
and skills required for a specific vacancy, identified by its \code{vacancy_id}. It constructs
and executes a SQL query to fetch the relevant data from the \code{adem.vacancies}, \code{adem.companies},
\code{adem.vacancy_skills}, and \code{adem.skills} tables.
}
\examples{
# Example of calling the function with a specific vacancy ID
result <- get_vacancy_by_id(vacancy_id = 123)

# Example of calling the function without any specific ID, which retrieves all vacancies
all_vacancies <- get_vacancy_by_id()

}
