% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vacancies.R
\name{get_vacancy_by_id}
\alias{get_vacancy_by_id}
\title{Obtains the details of a vacancy, given its ID, and the skills associated to it.}
\usage{
get_vacancy_by_id(vacancy_id = NULL)
}
\arguments{
\item{vacancy_id}{The ID of the vacancy whose details are to be retrieved.}
}
\value{
A list of 2 data frames: first, contains the vacancy details; second, contains list of associated skills. In case of an invalid or not found vacancy ID, the function returns NULL.
}
\description{
Obtains the details of a vacancy, given its ID, and the skills associated to it.
}
\examples{
\dontrun{
get_vacancy_by_id(123456)
}
}
