% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_vacancy_by_id.R
\name{get_vacancy_by_id}
\alias{get_vacancy_by_id}
\title{get_vacancy_by_id}
\usage{
get_vacancy_by_id(vacancy_id)
}
\arguments{
\item{vacancy_id}{integer}
}
\value{
list(vacancy, skills)
}
\description{
Returns full vacancy info + required skills. Returns NULL if not found.
}
\examples{
\dontrun{
get_vacancy_by_id(123456)
}
}
