% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_search.R
\name{log_search}
\alias{log_search}
\title{Log a user search query}
\usage{
log_search(user_id, query)
}
\arguments{
\item{user_id}{The unique identifier of the user.}

\item{query}{The search query string to log.}
}
\value{
TRUE if the log was successful, FALSE otherwise.
}
\description{
Log a user search query
}
