% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_logs.R
\name{log_search}
\alias{log_search}
\title{Insert search log}
\usage{
log_search(user_id = NULL, query)
}
\arguments{
\item{user_id}{Integer. Optional insert the user_id.}

\item{query}{Character. Query to be posted in the search log.}
}
\value{
A message confirming the action.
}
\description{
Insert your query in the search log.
}
\examples{
\dontrun{
# Inserts a query
log_search(query="Query test")

# Inserts a query with identified user_id
log_search(user_id=4, query="Second test")
}
}
