% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_logs.R
\name{log_search}
\alias{log_search}
\title{Log a search query from an API user}
\usage{
log_search(user_id, query)
}
\arguments{
\item{user_id}{Integer ID of the user}

\item{query}{Search query string}
}
\value{
TRUE if the log was inserted, FALSE otherwise
}
\description{
Log a search query from an API user
}
