% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_search.R
\name{log_search}
\alias{log_search}
\title{Logging user search queries in the database.}
\usage{
log_search(user_id = NULL, query = NULL)
}
\arguments{
\item{user_id}{A numeric type user ID.}

\item{query}{The search string entered by the user.}
}
\value{
Nothing.
}
\description{
Logging user search queries in the database.
}
\examples{
\dontrun{
log_search(12, "Python jobs")
log_search(user_id = 12, "Python jobs")
log_search(user_id = 12, query = "Python jobs")
}
}
