% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logs.R
\name{log_search}
\alias{log_search}
\title{Logs a search query in a specific schema table, given a user ID and a search query.}
\usage{
log_search(user_id, query)
}
\arguments{
\item{user_id}{The ID (an integer) of the user who issued the search query.}

\item{query}{The query string used in the search.}
}
\value{
TRUE, if logging was completed / successful; FALSE, if any error has occurred (e.g., invalid or not found user ID, invalid search query), preventing the logging.
}
\description{
Logs a search query in a specific schema table, given a user ID and a search query.
}
\examples{
\dontrun{
log_search(user_id = 1, query = "machine learning jobs")
}
}
