% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_logs.R
\name{log_search}
\alias{log_search}
\title{Log a User Search Query}
\usage{
log_search(user_id, query)
}
\arguments{
\item{user_id}{Integer. The user’s unique ID.}

\item{query}{Character. The text of the query to log.}
}
\value{
Logical value. TRUE if logging was successful, FALSE otherwise.
}
\description{
Records a search query made by a user into the `adem.search_logs` table.
}
\examples{
\dontrun{
log_search(user_id = 1, query = "machine learning jobs in Luxembourg")
}
}
