% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_search.R
\name{log_search}
\alias{log_search}
\title{log_search}
\usage{
log_search(user_id, query)
}
\arguments{
\item{user_id}{integer}

\item{query}{character}
}
\value{
TRUE/FALSE
}
\description{
Logs a search query in adem.search_logs. Returns TRUE on success, FALSE otherwise.
}
\examples{
\dontrun{
log_search(user_id = 1, query = "machine learning jobs")
}
}
