% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_search.R
\name{log_search}
\alias{log_search}
\title{Title}
\usage{
log_search(user_id, query)
}
\arguments{
\item{query}{}
}
\description{
Title
}
\examples{
\dontrun{
log_search(user_id = 1, query = "machine learning jobs")
}
}
