% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/token.R
\name{verify_token}
\alias{verify_token}
\title{Verify token}
\usage{
verify_token(token)
}
\arguments{
\item{token}{Character. Bearer token to access the database}
}
\value{
Boolean TRUE or FALSE
}
\description{
Verifies if the token is present in the database.
}
\examples{
\dontrun{
verify_token("testtoken")
}
}
