% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_token.R
\name{verify_token}
\alias{verify_token}
\title{Verify a bearer token}
\usage{
verify_token(token, schema = "adem")
}
\arguments{
\item{token}{A character string representing the bearer token to validate.
This parameter is required.}

\item{schema}{A character string representing the schema where the `api_users` table is located.
Defaults to `'adem'`.}
}
\value{
`TRUE` if the token is found in the database, `FALSE` otherwise.
}
\description{
Checks if a given bearer token exists in the `api_users` table of a specified schema.
}
